added remappings file and corrected 42_vyperJumpTables.t.sol

+++ test/26_staticSubcall.t.sol
@@ -1,11 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.15;
+
 import "forge-std/Test.sol";
 
 contract Foo {
     uint256 x;
 
-    function do_sstore() public returns (uint) {
+    function do_sstore() public returns (uint256) {
         unchecked {
             x += 1;
         }
@@ -13,7 +14,7 @@ contract Foo {
     }
 
     function test_staticcall_actually_writes() public view returns (bool success) {
-        (success, ) = address(this).staticcall(abi.encodeWithSignature("do_sstore()"));
+        (success,) = address(this).staticcall(abi.encodeWithSignature("do_sstore()"));
         assert(success);
     }
 }
