+++ test/28_staticContextPropagated.t.sol
@@ -4,13 +4,13 @@ pragma solidity ^0.8.15;
 contract Foo {
     event FooEvent();
 
-    function logFoo() public returns (uint) {
+    function logFoo() public returns (uint256) {
         emit FooEvent();
         return 42;
     }
 
     function view_func() public returns (bool succ) {
-        (succ, ) = address(this).call(abi.encodeWithSignature("logFoo()"));
+        (succ,) = address(this).call(abi.encodeWithSignature("logFoo()"));
     }
 
     function testStaticContext() public view {
